drm: xlnx: hdmi: fix 32-bit overflow in TMDS clock computation
xlnx_hdmi_get_tmdsclk() computes the TMDS clock as
"adjusted_mode->clock * 1000". adjusted_mode->clock is an int in kHz, so
the multiply is done in 32-bit signed arithmetic before the result is
widened into the u64 tmdsclk. xlnx_hdmi_connector_mode_valid() only
rejects modes above HDMI_TX_PIXEL_MAXRATE (2376000 kHz), so any clock
above 2147483 kHz overflows int. That is undefined behaviour and in
practice wraps negative and then sign-extends into a very large u64.

It is reachable with modes this driver advertises: 7680x4320p60 over FRL
sits at 2376000 kHz, exactly the advertised ceiling.

Cast adjusted_mode->clock to u64 before the multiply so the computation
is performed in 64-bit. Fixes Coverity OVERFLOW_BEFORE_WIDEN.

Fixes: c4cb4c415047 ("drm: xlnx: hdmi: Add HDMI 2.1 Tx subsystem driver")
Message-ID: <20260903063348.3118021-7-harish.bommisetti@amd.com>
State: pending

+++ drivers/gpu/drm/xlnx/xlnx_hdmi.c
@@ -3356,7 +3356,7 @@ static u64 xlnx_hdmi_get_tmdsclk(struct xlnx_hdmi *hdmi, struct drm_display_mode
 {
 	u64 tmdsclk;
 
-	tmdsclk = adjusted_mode->clock * 1000;
+	tmdsclk = (u64)adjusted_mode->clock * 1000;
 
 	if ((adjusted_mode->flags & DRM_MODE_FLAG_DBLCLK) &&
 	    (adjusted_mode->flags & DRM_MODE_FLAG_INTERLACE))
